[revert] revert 0e4de28694ced2498e3caa26de5427c9a2d51d7c to reimplement forkJoin
There are two options for fetching all the calls to `lineage-mutations`:
1. Run a forkJoin over all the queries and loop over each `pangolin_lineage` to fetch results and merge together(currently what's running on the main branch).
2. Combine all the `pangolin_lineages` into a single comma-separated query for the backend, taking advantage of the looping built into the API endpoint.

Neither option is perfect

The second option is faster for an individual query. However, especially for VOCs which are expanded into a series of `lineage OR lineage` queries, you quickly run into URI too long errors. Also, if you sequentially add lineages, you can't take advantage of NGINX caching, since each addition changes the API request.

In addition to the speed issue with the first method, though, the other issue is that if you request data for a lot of lineages, you run into throttling/429 errors, because you're requesting more data.

We should revisit this when:
* we have a more robust way of querying lineages and sublineages.
* the lineages-mutations endpoint is optimized to be faster.

If it's fast and the queries have fewer characters, we'll probably want to move back to using the API to execute one query instead of a `forkJoin` loop of queries.

We might also want to alert the user they've selected too many lineages and/or add a pause between groups of queries if they've selected too many.